requirements.txt
get==1.0.3
numpy==1.15.2
pandas==0.23.4
Pillow==5.2.0
post==1.0.2
public==1.0.3
python-dateutil==2.7.3
pytz==2018.5
query-string==1.0.2
request==1.0.2
scikit-learn==0.20.0
scipy==1.1.0
six==1.11.0
torch==0.4.1
torchvision==0.2.1
